fix: download model even if train.py is called with pre-existing JSON configuration.
