+++ picotron/utils.py
@@ -1,8 +1,14 @@
-import torch
+import json
+import os
+import subprocess
+import requests
 import random
 import numpy as np
 import builtins
 import fcntl
+
+from safetensors import safe_open
+
 import picotron.process_group_manager as pgm
 import torch, torch.distributed as dist
 
@@ -93,3 +99,114 @@ def average_loss_across_dp_cp_ranks(loss, device):
         dist.all_reduce(reduced_loss, op=dist.ReduceOp.SUM, group=pgm.process_group_manager.cp_dp_group)
         reduced_loss /= pgm.process_group_manager.cp_dp_world_size
     return reduced_loss.item()
+
+def check_hf_model_files_existences(model_name, hf_token):
+    files_to_check = [
+        "model.safetensors",
+        "model.safetensors.index.json"
+    ]
+
+    # Prepare headers with authentication token
+    headers = {}
+    if hf_token: headers["Authorization"] = f"Bearer {hf_token}"
+
+    index = 0
+    found_files = []
+    for file in files_to_check:
+        url = f'https://huggingface.co/{model_name}/resolve/main/{file}'
+        try:
+            # Use GET request with stream=True and authentication headers
+            response = requests.get(url, stream=True, headers=headers)
+            if response.status_code == 200:
+                found_files.append(file)
+                print(f"✅ Found {file}")
+                response.close()
+            elif response.status_code == 401:
+                print(f"❌ Authentication required for {file} (Status: {response.status_code})")
+            elif response.status_code == 403:
+                print(f"❌ Access denied for {file} (Status: {response.status_code})")
+            else:
+                print(f"❌ Not found {file} (Status: {response.status_code})")
+        except Exception as e:
+            print(f"❌ Error checking {file}: {str(e)}")
+
+    return found_files
+
+def download_hf_model_files(files_to_download, model_name, hf_token, save_dir):
+    downloaded_files = []
+
+    save_dir_path = f"{save_dir}/{model_name}"
+
+    for file in files_to_download:
+        if os.path.exists(os.path.join(save_dir_path, file)):
+            print(f"✅ {file} already exists")
+            downloaded_files.append(file)
+
+            # If it's index.json, read it to get shards
+            if file.endswith('.json'):
+                with open(os.path.join(save_dir_path, file), 'r') as f:
+                    index_data = json.load(f)
+                    shards = set(index_data['weight_map'].values())
+                    print(f"Found {len(shards)} shards in index")
+                    files_to_download.extend(shards)
+            continue
+
+        model_cmd = f"huggingface-cli download {model_name} {file} --local-dir {save_dir_path} --token {hf_token}"
+        print(f"Downloading {file}...")
+        env = os.environ.copy()
+        env["HF_HUB_ENABLE_HF_TRANSFER"] = "1"
+        env["PYTHONUNBUFFERED"] = "1"
+        result = subprocess.run(model_cmd, shell=True, check=False, env=env)
+
+        os.environ["HF_HUB_ENABLE_HF_TRANSFER"] = "1"
+
+        if result.returncode == 0:
+            print(f"✅ {file} downloaded successfully")
+            downloaded_files.append(file)
+
+            # Verify files based on their type
+            file_path = os.path.join(save_dir_path, file)
+            if file.endswith('.safetensors'):
+                try:
+                    with safe_open(file_path, framework="pytorch", device="cpu") as f:
+                        keys = list(f.keys())
+                        print(f"✅ Safetensors file is valid")
+                        print(f"- Number of tensors: {len(keys)}")
+                except Exception as e:
+                    print(f"❌ Error validating safetensors file: {str(e)}")
+                    continue
+            elif file.endswith('.json'):
+                try:
+                    with open(file_path, 'r') as f:
+                        index_data = json.load(f)
+                        shards = set(index_data['weight_map'].values())
+                        print(f"✅ Index JSON file is valid")
+                        print(f"- Number of weight shards: {len(shards)}")
+                        # Add shards to files_to_download
+                        files_to_download.extend(shards)
+                except Exception as e:
+                    print(f"❌ Error validating index JSON file: {str(e)}")
+                    continue
+        else:
+            error_message = result.stderr.decode('utf-8', errors='replace')
+            if "404 Client Error" in error_message or "Entry Not Found" in error_message:
+                print(f"❌ File {file} not found in repository")
+            else:
+                print(f"❌ Download failed: {error_message.strip()}")
+
+    print(f"\nSuccessfully downloaded files: {', '.join(downloaded_files)}")
+    return True
+
+def download_model(model_name, hf_token):
+    # Download HF model safetensors at the "hf_model_safetensors" directory
+    os.makedirs("hf_model_safetensors", exist_ok=True)
+
+    files_to_download = check_hf_model_files_existences(model_name, hf_token)
+    if len(files_to_download) <= 0:
+        raise FileNotFoundError("Safetensors files not found. Please check the model name and authentication token.")
+
+    is_downloaded = download_hf_model_files(files_to_download, model_name, hf_token, save_dir="hf_model_safetensors")
+    if not is_downloaded:
+        raise FileNotFoundError("Failed to download safetensors files. Please check the model name and authentication token.")
+
+    print("SafeTensors files downloaded successfully! ✅")
